Refactoring list action and get_user_by_id
In order to solve ticket #52 I felt I wanted to refactor the existing
list action and the get_user_by_id.

The changes to get_user_by_id is mostly use return instead of yield.
Before we mixed both return and yield which is not a good thing to do.
Also, there was no point of yielding when we only call this function
once per list action.

Added initial logic for list action to support "today".
Added test for get_user_by_id.