Use SQS for commands
We're seeing random timeouts when waiting for response to slack. This
replaces the previous sync response with a quick ack that the message
has been received and then handle it in the background.

Since chalice doesn't create the queues automatically it's done from
travis.

NOTE: We need to do the first deploy locally, as travis doesn't save the
required chalice state

Fix #191
Fix #99